libheif_rs::regions

Enum Region

#[non_exhaustive]
pub enum Region {
    Point(PointRegion),
    Rectangle(RectangleRegion),
    Ellipse(EllipseRegion),
    Polygon(PolygonRegion),
    Polyline(PolylineRegion),
    ReferencedMask(ReferencedMaskRegion),
    InlineMask(InlineMaskRegion),
}

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Point(PointRegion)

Point geometry.

The region is represented by a single point.

Rectangle(RectangleRegion)

Rectangle geometry.

The region is represented by a top left position, and a size defined by a width and height. All the interior points and the edge are part of the region.

Ellipse(EllipseRegion)

Ellipse geometry.

The region is represented by a center point, and radii in the X and Y directions. All the interior points and the edge are part of the region.

Polygon(PolygonRegion)

Polygon geometry.

The region is represented by a sequence of points, which is considered implicitly closed. All the interior points and the edge are part of the region.

Polyline(PolylineRegion)

Polyline geometry.

The region is represented by a sequence of points, which are not considered to form a closed surface. Only the edge is part of the region.

ReferencedMask(ReferencedMaskRegion)

Referenced mask.

The mask location is represented by a left-top corner position, and a size defined by a width and height. The value of each sample in that mask identifies whether the corresponding pixel is part of the region.

The region geometry is described by the pixels in another image item, which has an item reference of type mask from the region item to the image item containing the mask.

The image item containing the mask is one of:

a mask item (see ISO/IEC 23008-12:2022 Section 6.10.2), or a derived image from a mask item;

an image item in monochrome format (4:0:0 chroma);

an image item in format of color with luma and chroma planes (e.g. 4:2:0).

If the pixel value is equal to the minimum sample value (e.g. 0 for unsigned integer), the pixel is not part of the region. If the pixel value is equal to the maximum sample value (e.g. 255 for 8 bit unsigned integer), the pixel is part of the region. If the pixel value is between the minimum sample value and maximum sample value, the pixel value represents an (application defined) probability that the pixel is part of the region, where higher pixel values correspond to higher probability values.

InlineMask(InlineMaskRegion)

Inline mask.

The region geometry is described by a sequence of bits stored in inline in the region, one bit per pixel. If the bit value is 1, the pixel is part of the region. If the bit value is 0, the pixel is not part of the region.