libheif_rs

Trait Reader

pub trait Reader {
    // Required methods
    fn position(&mut self) -> u64;
    fn read(&mut self, buf: &mut [u8]) -> Result<usize>;
    fn seek(&mut self, position: u64) -> Result<u64>;
    fn wait_for_file_size(&mut self, target_size: u64) -> ReaderGrowStatus;
}

Required Methods

fn position(&mut self) -> u64

Current position, in bytes, inside of a source.

fn read(&mut self, buf: &mut [u8]) -> Result<usize>

Pull some bytes from a source into the specified buffer, returning how many bytes were read.

fn seek(&mut self, position: u64) -> Result<u64>

Seek to an position, in bytes, from start of a source.

fn wait_for_file_size(&mut self, target_size: u64) -> ReaderGrowStatus

Wait until a source will be ready to read bytes to the specified position.